+++ src/PhpWord/Element/Row.php
@@ -50,14 +50,11 @@ class Row extends AbstractElement
     /**
      * Create a new table row
      *
-     * @param string $docPart
-     * @param int $docPartId
      * @param int $height
      * @param mixed $style
      */
-    public function __construct($docPart, $docPartId, $height = null, $style = null)
+    public function __construct($height = null, $style = null)
     {
-        $this->setDocPart($docPart, $docPartId);
         $this->height = $height;
         $this->style = $this->setStyle(new RowStyle(), $style, true);
     }
@@ -67,12 +64,15 @@ class Row extends AbstractElement
      *
      * @param int $width
      * @param mixed $style
+     * @return \PhpOffice\PhpWord\Element\Cell
      */
     public function addCell($width = null, $style = null)
     {
-        $cell = new Cell($this->getDocPart(), $this->getDocPartId(), $width, $style);
+        $cell = new Cell($width, $style);
+        $cell->setDocPart($this->getDocPart(), $this->getDocPartId());
         $cell->setPhpWord($this->phpWord);
         $this->cells[] = $cell;
+
         return $cell;
     }