+++ src/PhpWord/Element/Section.php
@@ -27,6 +27,11 @@ use PhpOffice\PhpWord\Style\Section as SectionSettings;
 class Section extends AbstractContainer
 {
     /**
+     * @var string Container type
+     */
+    protected $container = 'Section';
+
+    /**
      * Section settings
      *
      * @var \PhpOffice\PhpWord\Style\Section
@@ -55,7 +60,6 @@ class Section extends AbstractContainer
      */
     public function __construct($sectionCount, $settings = null)
     {
-        $this->container = 'section';
         $this->sectionId = $sectionCount;
         $this->setDocPart($this->container, $this->sectionId);
         $this->settings = new SectionSettings();
@@ -98,15 +102,7 @@ class Section extends AbstractContainer
      */
     public function addTitle($text, $depth = 1)
     {
-        $title = new Title($text, $depth);
-        $title->setDocPart($this->getDocPart(), $this->getDocPartId());
-        if ($this->phpWord instanceof PhpWord) {
-            $bookmarkId = $this->phpWord->addTitle($title);
-            $title->setBookmarkId($bookmarkId);
-        }
-        $this->addElement($title);
-
-        return $title;
+        return $this->addElement('Title', $text, $depth);
     }
 
     /**
@@ -114,7 +110,7 @@ class Section extends AbstractContainer
      */
     public function addPageBreak()
     {
-        $this->addElement(new PageBreak());
+        return $this->addElement('PageBreak');
     }
 
     /**
@@ -128,10 +124,7 @@ class Section extends AbstractContainer
      */
     public function addTOC($fontStyle = null, $tocStyle = null, $minDepth = 1, $maxDepth = 9)
     {
-        $toc = new TOC($fontStyle, $tocStyle, $minDepth, $maxDepth);
-        $this->addElement($toc);
-
-        return $toc;
+        return $this->addElement('TOC', $fontStyle, $tocStyle, $minDepth, $maxDepth);
     }
 
     /**